Rename KeyBit file tree
