+++ src/types/MapObject.ts
@@ -1,4 +1,4 @@
-export type MapObject = {
+type MapObject = {
   mapName: string;
   mapSrc: string;
   textures: Texture[];
@@ -19,3 +19,5 @@ export type InteractiveCell = {
   interaction: string;
   cell: Cell;
 };
+
+export default MapObject;
